important changes, see description
to run the program you just do ./test dataoutput.txt settings.txt
no need for the path of the input file instead that should be put in
line two of settings. If that is set to null it will default and look
for it in the current dir. Also some checks to make sure that it is
opening properly.
Also added the function for usr1 LED to “blink” it works for me but
doesn’t really blink it just stays solid so going to look into that
some more. Also changing the number of days in line one of settings
works but is capped at 120 for data allocation. The only thing that
things to be fixed there is your puts to the file, not sure how you
were doing that so didn’t want to touch it.